Comprehensive codebase review and admin interface improvements
Backend fixes:
- Remove duplicate import in firebase.py
- Fix type annotations and missing import in customer.py
- Optimize integer column type in plan.py
- Add JWT payload validation in auth deps
- Add query limit optimization in drivers endpoint
- Consolidate test utilities to reduce code duplication

Frontend improvements:
- Implement dual driver support in routes interface
- Convert inline styles to consistent Tailwind CSS classes
- Enhance admin page styling with cards, loading states, and responsive design
- Improve modal components with better UX and loading indicators
- Add environment-aware console logging
- Update table styling with hover effects and better typography

Android fixes:
- Fix missing icon imports causing build failures
- Add authentication to FCM token uploads
- Clean up unused imports and dead code

+++ backend/app/models/plan.py
@@ -2,7 +2,7 @@ from decimal import Decimal
 
 from datetime import date
 
-from sqlalchemy import BigInteger, Date, String, Numeric, ForeignKey
+from sqlalchemy import BigInteger, Date, String, Numeric, Integer, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column
 
 from .base import Base
@@ -13,7 +13,7 @@ class Plan(Base):
     order_id: Mapped[int] = mapped_column(ForeignKey("orders.id"), nullable=False)
     plan_type: Mapped[str] = mapped_column(String(20))  # RENTAL | INSTALLMENT
     start_date: Mapped[date | None] = mapped_column(Date, nullable=True)
-    months: Mapped[int | None] = mapped_column(Numeric(12, 0), nullable=True)  # For installment
+    months: Mapped[int | None] = mapped_column(Integer, nullable=True)  # For installment
     monthly_amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), default=0)
     upfront_billed_amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), default=0)
     status: Mapped[str] = mapped_column(String(20), default="ACTIVE")  # ACTIVE|CANCELLED|COMPLETED
